1.5 is completed, although I might need to test some more things.
But for now, it is completed. Didn't think I'd be able to do it, but I did. Apparently there was something wrong with Atom()'s evaluate function, but I fixed it, and I could just go on to 1.6, but 1.6 seems to be impossible, as I'm not sure if I'll be able to finish putting expressions in CNF form...