After some trouble, fiddling, and Googling, I managed to solve the first part.
Now as it turns out, it seems that the professor neglected to mention that it was necessary to use open() on the given email_path before using the email.message_from_file() method. I spent 30 minutes trying to figure this one out and only found out the answer after googling and finding out why the method wouldn't work. Since I kept getting an error on "str does not have .read()" or something like that, I had my suspicions, but I assumed that the method would have automatically opened it.

Well, it doesn't. Anyway, once I figured that part out, it didn't take long for me to figure the rest out. Ech.