Turns out, I swapped the values for Sigma count(w') and |V|.
I guess I mixed up the two. I probably misunderstood the equation and the information I gpt from talking to some of my classmates... In any case, all the test cases for question 2 pass with flying colors, so that's something to be grateful for. On to question 3, assuming that I haven't failed the private tests for question 2.