Fixed a bug that caused errors on Gradescope.
I had to fix a bug in load_corpus() that prevented the code from running on Gradescope, as it would keep on saying that the autograder failed to execute correctly. It turns out, .readlines() is not a good thing to use for Gradescope. The original code worked fine when I tested it on my computer, but Gradescope didn't seem to be capable of running it. Anyway, the new implementation works like the original, and it also runs on Gradescope, so that's what this edit was for.